deluxe

Macro define_with_optional

macro_rules! define_with_optional {
    (
        $(#[$attrs:meta])* $vis:vis mod $mod:ident,
        $($path:ident)? $(:: $path_rest:ident)*,
        $ty:ty $(,)?
    ) => { ... };
}

Generates a module for parsing an optional value using `#[deluxe(with = …)].

Takes three arguments separated by commas:

The generated module. Can include attributes and a visibility specifier.

The module to use as an inner parser, relative to the new module.

The target type.

Example

Defines a new module named mod_path_optional that parses an Option using with::mod_path.

deluxe_core::define_with_optional!(
    pub mod mod_path_optional,
    deluxe_core::with::mod_path,
    syn::Path
);