deluxe

Struct Errors

pub struct Errors { /* private fields */ }

A wrapper for a list of errors. Can be empty.

Implementations

impl Errors

pub const fn new() -> Errors

Creates a new empty error list.

pub fn is_empty(&self) -> bool

Checks if the list contains any errors.

pub fn clear(&self)

Reset the list to an empty state.

pub fn push<T>(&self, span: Span, message: T)
where T: Display,

Pushes one error onto the list. This function is a wrapper around syn::Error::new.

pub fn push_call_site<T>(&self, message: T)
where T: Display,

Pushes one error onto the list, setting the errorโs span to Span::call_site().

pub fn push_spanned<T, U>(&self, tokens: T, message: U)
where T: ToTokens, U: Display,

Pushes one error onto the list spanning the given syntax tree node. This function is a wrapper around syn::Error::new_spanned.

pub fn push_syn(&self, error: Error)

Pushes one previously constructed Error onto the list.

pub fn push_result<T>(&self, result: Result<T, Error>) -> Option<T>

Pushes an error onto the list from a Result.

If result is Err, pushes the error and returns None. If result is Ok, returns Some(T).

pub fn extend<T>(&self, iter: T)
where T: IntoIterator<Item = Error>,

Appends all errors from iter into this list.

pub fn into_result<T>(self, value: T) -> Result<T, Error>

Returns Err if the list has errors, or Ok(value) if the list is empty.

If the list has any errors, returns Err containing one Error with all of the errors combined using Error::combine.

pub fn check(self) -> Result<(), Error>

Checks if the error list is empty.

If the list has any errors, returns Err containing one Error with all of the errors combined using Error::combine. Otherwise, returns Ok.

pub fn unwrap_err(self) -> Error

Returns the inner if the error list has errors.

Panics

Panics if the error list is empty.

pub fn bail<T>(self) -> Result<T, Error>

Returns a new Err if the error list has errors.

Panics

Panics if the error list is empty.

pub fn into_compile_error(self) -> Option<TokenStream>

Converts the error list into a token stream containing std::compile_error invocations.

The errors are generated with Error::into_compile_error.

Returns None if the list is empty.

pub fn into_compile_errors(self) -> impl IntoIterator<Item = TokenStream>

Returns an iterator of token streams containing std::compile_error invocations.

Each token stream will contain one invocation. The errors are generated with Error::into_compile_error.

pub fn output_with<Q>(self, item: Q) -> TokenStream
where Q: ToTokens,

Creates a token stream containing the current set of errors and item.