deluxe::with
Module maybe_quoted

Helpers for parsing any type that implements ParseMetaItem possibly parsed out of a quoted string first.

Can be used on a field by specifying the module, like #[deluxe(with = deluxe::with::maybe_quoted)]

Functions

missing_meta_item

parse_meta_item

parse_meta_item_flag

parse_meta_item_inline

parse_meta_item_named