deluxe

Function extract_attributes_optional

pub fn extract_attributes_optional<T, R>(obj: &mut T, errors: &Errors) -> R
where
    T: HasAttributes,
    R: ExtractAttributes<T> + Default,

Extracts attributes out of another type holding a list of syn::Attribute, then parses them into a Rust type, returning a default value on failure.

Calls extract_attributes and returns the result. Upon failure, Default::default will be returned and any errors will be appended to errors. This function should be preferred when parsing multiple attributes, so the errors from all of them can be accumulated instead of returning early.